dom/media/tests/mochitest/TextTrack_kind.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
  <title>Test for Bug 833386</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=833386">Mozilla Bug 833386</a>
<p id="display"></p>
<div id="content" style="display: none">
  <video id="video_1">
    <track id="track_1" src="fast.vtt" />
  </video>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
  SpecialPowers.setBoolPref("media.webvtt.enabled", true);
  SimpleTest.waitForExplicitFinish();
  
  var trackElement = document.getElementById('track_1');
  var textTrack = trackElement.track;

  function setKind(kind) {
    // Test via dom
    trackElement.kind = kind;
    is(textTrack.kind, kind);
  }

  function start() {
    // Test Defaults
    is(textTrack.kind, "subtitles");

    // Test Bogus via dom
    trackElement.kind = 'bogus';
    is(textTrack.kind, "subtitles");

    // set to known values
    setKind('subtitles');
    setKind('captions');
    setKind("descriptions");
    setKind("chapters");
    setKind("metadata");

    SimpleTest.finish();
  }
  addLoadEvent(start);
</script>
</pre>
</body>
</html>
